Add MethodFilterExtensionsTests
